requirements.txt
# requirements.txt (prod)
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
asterisk-ari==1.0.1
websockets>=11.0
python-dotenv>=1.0.0
numpy>=1.24.0
soundfile>=0.12.1
pytest>=7.0.0